# evmrs: use push guard to avoid stack overflow checks