--- a/rust/src/types/stack.rs
+++ b/rust/src/types/stack.rs
@@ -127,24 +127,18 @@ impl Stack {
     pub fn pop_with_guard<const N: usize>(&mut self) -> Result<(PushGuard, [u256; N]), FailStatus> {
         self.check_underflow(N)?;
 
-        unsafe {
-            self.0.set_len(self.len() - N + 1);
-        }
-        let start = self.0.as_ptr() as *mut u256;
+        self.0.truncate(self.len() - (N - 1));
         // SAFETY:
         // This does not wrap and the whole range from start to start + self.len is valid.
-        let pop_start = unsafe { start.add(self.len() - 1) };
+        let pop_start = unsafe { self.0.as_ptr().add(self.len() - 1) };
         // SAFETY:
         // The the first self.len elements are initialized (invariant).
-        // `self.len` just got decremented by N - 1, which means now that the first `self.len - 1 +
-        // (N + 1)` elements are initialized. Therefore, it is safe to read N elements
+        // `self.len` just got decremented by N - 1, which means now that the first `self.len  +
+        // (N - 1)` elements are initialized. Therefore, it is safe to read N elements
         // starting at index `self.len - 1` as an array of length N and type u256.
         let pop_data = unsafe { *(pop_start as *const [u256; N]) };
-        // SAFETY:
-        // The data for pop_data is copied out so there are no other references to this data.
-        // The validity of the data is the same as for pop_data. Because the pointer is valid and no
-        // one else holds a reference to it, it is safe to cast it to a mutable reference.
-        let push_guard = PushGuard(unsafe { &mut *pop_start });
+        let len = self.len();
+        let push_guard = PushGuard(&mut self.0[len - 1]);
         Ok((push_guard, pop_data))
     }